backend/requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
cryptography==46.0.6
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.128.8
frozenlist==1.8.0
google-auth==2.49.1
google-auth-oauthlib==1.3.0
gspread==6.2.1
h11==0.16.0
idna==3.11
magic-filter==1.0.12
multidict==6.7.1
numpy==2.0.2
oauthlib==3.3.1
openpyxl==3.1.5
pandas==2.3.3
propcache==0.4.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2026.1.post1
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
sse-starlette==3.3.0
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.39.0
yarl==1.22.0
